[Avidemux] Sync with main development tree

+++ Avidemux/plugins/ADM_videoFilters/Srt/ADM_vidSRTload.cpp
@@ -394,7 +394,7 @@ int ADM_SubAtoi(uint16_t *in)
 uint8_t	ADM_utfInit( char *charset )
 {
 	myConv=iconv_open("UTF-16",charset); //"WINDOWS-1251");
-	if((long int)myConv==-1)
+	if((intptr_t)myConv==-1)
 	{
 		printf("\n Error initializing iconv...\n");
 		return  0;
@@ -474,7 +474,7 @@ uint32_t done=0;
 uint8_t ADM_utfEnd( void)
 {
   // Purge iconv
-  if((long int)myConv!=-1)
+  if((intptr_t)myConv!=-1)
 	{
 		iconv_close(myConv);
 		myConv=(iconv_t)-1;
